Allow Arbitrary Metadata for All Content
This change adds a `meta` attribute to the Front Matter section of every type of content supported by `empress-blog`. This attribute is treated as an arbitrary JSON blob by the Ember-Data models, supporting string, array, and/or object formats.

The `meta` attribute is not used by `empress-blog` by default, but it enables developer support for adding metadata to authors, content, pages, and tags.

This is currently undocumented as it is an expert feature