Use actual local image embed when mentioned in guide content
Neither the demo app nor the blueprint utilized an actual local image embed despite talking specifically about it in the guides and purporting to show an example. This updates the example to use an image from the public/ folder rather than an externally-hosted image. There is already an example of an externally-hosted image in one of the other content files, so I left that one as an example. It seemed appropriate to change this one since it's where the guides specifically talk about using images in the public/ directory in empress-blog